Extract table component
